Fix state when editing existing galleries
Previously, when opening an existing gallery for editing, the user
would be presented with the add new/upload media view and a few
seconds later, the associated gallery images would load. There was
no indication that anything was loading and the overall experience
was kind of weird. By setting the state to "gallery-library", when
the media window opens, a loading indicator is displayed until the
images are fetched via ajax.